requirements.txt
pip
python-dotenv
py-cord==2.2.2